Update Dockerfile
Update the Dockerfile to reflect what we're running in production.
The Dockerfile does not match what's used when we're building the image
on servers.

I'll be using this Dockerfile when performing automatic deployments.

There only noticeable change I've is that I'm using Node 8 LTS instead
of Node 9. I think for the purpose of stability we should be using LTS.